RA-1082 Combine and Fix Registration Tests
The two tests were doing the same thing, so they were combined.
Additionally, the tests now conform to the automated testing
guidelines.
Also, pom.xml has been updated with the latest selenium dependency, so
that the tests will run with the newest version of Firefox. However,
the tests must be executed with Java 8 to work with new Selenium. This
is a temporary fix until the dependency is added to all of OpenMRS.
I’ve already submitted a pull request on the core to have that happen.