O3-1756 dependency: Adds required codes for basic HPI & SOAP forms
Request from Faimer was for demo HPI (History of Present Illness) & SOAP (Sub/Obj/Assess/Plan) form templates to be added to O3 Demo. We agreed these are actually very valuable to have in the RefApp in general. This PR adds concepts required for these forms. 
This is required for O3-1756.