O3-2047: Add 10 more sample locations to help catch >50 loading issues in future
If I understand https://github.com/openmrs/openmrs-esm-patient-chart/pull/1131 correctly, it didn't help that we didn't have >50 sample location sites in the refapp for testing purposes. I actually thought that I had 100+ sample locations in the config for this reason, but we only had 50. This PR just adds another 10 in the hopes this is helpful. 
If this is not helpful, let me know.