O3-2509: Adding a TestsOrderability ConvSet Concept
This update to the Basic Lab Tests collection includes a custom openmrs concept we are using to organize labs that we want to show up as orderable options. This is a first version with limited content examples but should help us unblock the dev work and confirm the approach so we can move forward.